changes to JEC JASPER-219 +review JSB @awehbi
made changes so JEC has access to connector to get the URI from JTA. This allows them to make uri anything they want. Made changes to delegate to allow for multiple queues per URI and had to add new map to track JTA name against queue because when connection to JTA is lost, broker only knows JTA name. So we need to use this to find queues so we can cleanup uri map in delegate