fix for JASPER-270
we should not encode/decode URIs. This would have to be done on JTA and JSC sides and adds complexity not required. It is a valid error scenario for there to be a mismatch from what is configured on JTA for URI and what comes in from JSC. We should just not find URI and log error and reply to JSC. This fix takes out the URI decoding and we store whatever comes in from JTA even if it is encoded