JASPER-273 #resolve #comment this is the deployable war file for tomcat which drive the client's end to end synchronous interaction. Can be provided to future customers for inter-operability demonstration and as a guidance to use our JClient. #time 1d +review JSB @mromaniuk @jmorales @awehbi @rouellette
All you need to deploy to tomcat in one spot.
once it's build, pick up the jasper-1.0.war from the target directory and drop it in Tomcat's webapp directory.  All dependencies are included in this bundle so you no longer need to copy the JClient jar to the tomcat lib dir. As a matter of fact, please remove it from the lib dir.