JASPER-645 GET for DTA-E with no data returns invalid response (200ok).  Fix also affects POST (which for success returns 200ok with null response) #resolve #time 1d +review JSB @awehbi @rouellette
GET and POST (in)directly invoke sendAndWaitForResponse which if it receives 200 ok and null text response, hard codes a {"code":200, reason:"ok"} and gives this back to processGetRequest and processPostRequest.  For POST this is ok, but not for GET (no data case).  Solution is to remove the hard coding of 200/ok from sendAndWaitForResponse, instead throw exception if we get null response and  status code is not 200.  Then let processGetRequest and processPostRequest decide how to handle null response since code should be 200.

Signed-off-by: Paul Granville <pgranville@coralcea.ca>